Flytt avbryt jobber til en fil
