+++ src/main/kotlin/dokumentkobling/AvbrytDokumentJobb.kt
@@ -3,10 +3,12 @@ package no.nav.helsearbeidsgiver.dokumentkobling
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import no.nav.hag.utils.bakgrunnsjobb.RecurringJob
+import no.nav.helsearbeidsgiver.database.DokumentkoblingRepository
 import java.time.Duration
 import java.time.LocalDateTime
 
 private const val ANTALL_MINUTTER_MELLOM_KJOERINGER = 1L
+private const val ANTALL_MINUTTER_FOER_TIDSAVBRUDD = 60L
 
 abstract class AvbrytDokumentJobb(
     private val dokumentNavn: String,
@@ -22,3 +24,35 @@ abstract class AvbrytDokumentJobb(
         logger.info("Satte $antallAvbrutte $dokumentNavn til status TIDSAVBRUTT")
     }
 }
+
+class AvbrytForespoerselJobb(
+    dokumentkoblingRepository: DokumentkoblingRepository,
+) : AvbrytDokumentJobb(
+        dokumentNavn = "forespørsler",
+        antallMinutterFoerTidsavbrudd = ANTALL_MINUTTER_FOER_TIDSAVBRUDD,
+        settTilTidsavbrutt = dokumentkoblingRepository::settForespoerslerMedStatusMottattTilTidsavbrutt,
+    )
+
+class AvbrytInntektsmeldingJobb(
+    dokumentkoblingRepository: DokumentkoblingRepository,
+) : AvbrytDokumentJobb(
+        dokumentNavn = "inntektsmeldinger",
+        antallMinutterFoerTidsavbrudd = ANTALL_MINUTTER_FOER_TIDSAVBRUDD,
+        settTilTidsavbrutt = dokumentkoblingRepository::settInntektsmeldingerMedStatusMottattTilTidsavbrutt,
+    )
+
+class AvbrytSykepengeSoeknadJobb(
+    dokumentkoblingRepository: DokumentkoblingRepository,
+) : AvbrytDokumentJobb(
+        dokumentNavn = "sykepengesøknader",
+        antallMinutterFoerTidsavbrudd = ANTALL_MINUTTER_FOER_TIDSAVBRUDD,
+        settTilTidsavbrutt = dokumentkoblingRepository::settSykepengeSoeknaderMedStatusMottattTilTidsavbrutt,
+    )
+
+class AvbrytSykmeldingJobb(
+    dokumentkoblingRepository: DokumentkoblingRepository,
+) : AvbrytDokumentJobb(
+        dokumentNavn = "sykmeldinger",
+        antallMinutterFoerTidsavbrudd = ANTALL_MINUTTER_FOER_TIDSAVBRUDD,
+        settTilTidsavbrutt = dokumentkoblingRepository::settSykmeldingerMedStatusMottattTilTidsavbrutt,
+    )
